ai-backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "ipywidgets>=8.1.8",
    "jq>=1.11.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.5",
    "lxml>=6.0.2",
    "psycopg2>=2.9.11",
    "pymongo[srv]>=4.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tavily>=1.1.0",
    "uvicorn>=0.40.0",
]
